refactor: :art: error handling DB
use only one type of error for crate db
