+++ crates/client/db/src/codec.rs
