+++ crates/client/sync/src/commitments/contracts.rs
@@ -4,8 +4,8 @@ use bitvec::order::Msb0;
 use bitvec::vec::BitVec;
 use bitvec::view::AsBits;
 use bonsai_trie::id::BasicId;
-use dc_db::storage_handler::{bonsai_identifier, DeoxysStorageError, StorageType};
 use dc_db::DeoxysBackend;
+use dc_db::{bonsai_identifier, DeoxysStorageError, StorageType};
 use dp_block::{BlockId, BlockTag};
 use dp_convert::ToFelt;
 use indexmap::IndexMap;
