+++ crates/client/sync/src/commitments/events.rs
@@ -2,7 +2,7 @@ use bitvec::vec::BitVec;
 use bonsai_trie::databases::HashMapDb;
 use bonsai_trie::id::{BasicId, BasicIdBuilder};
 use bonsai_trie::{BonsaiStorage, BonsaiStorageConfig};
-use dc_db::storage_handler::bonsai_identifier;
+use dc_db::bonsai_identifier;
 use dp_receipt::Event;
 use rayon::prelude::*;
 use starknet_types_core::felt::Felt;
